requirements.txt
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
idna==3.10
joblib==1.4.2
llvmlite==0.43.0
numba==0.60.0
numpy==2.0.2
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
pynndescent==0.5.13
python-dateutil==2.9.0.post0
python-multipart==0.0.10
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.6
uvicorn==0.30.6
